dec/vp8l: quiet vs9 x64 type conversion warning
src\dec\vp8l.c(816) : warning C4244: '=' : conversion from '__int64' to
'int', possible loss of data
src\dec\vp8l.c(817) : warning C4244: '=' : conversion from '__int64' to
'int', possible loss of data

+++ src/dec/vp8l.c
@@ -794,7 +794,7 @@ End:                                                                           \
     dec->status_ =                                                             \
         (!br->eos_) ? VP8_STATUS_BITSTREAM_ERROR : VP8_STATUS_SUSPENDED;       \
   } else {                                                                     \
-    dec->last_pixel_ = src - data;                                             \
+    dec->last_pixel_ = (int)(src - data);                                      \
     if (src == src_end) dec->state_ = READ_DATA;                               \
   }                                                                            \
   return ok;                                                                   \
